api/requirements.txt
Flask==3.1.0
google-generativeai==0.8.3
firebase-admin==6.5.0
python-dotenv==1.1.0
requests==2.32.3
